Refactor: Cycles: Make kernel_features 64bit
This change adds 32 more bit to store kernel features.

While for a short term it might be possible to make a space for one or
two extra bits, it seems going 64bit is inevitable.

Expanding the field to 64bit might introduce some slowdown due to less
optimal cache, but so is consolidation of existing flags could also
lead to performance drop in certain configurations.

The main tricky part of the change is Metal where function constants
are used to store kernel_features, and 64bit constants are only
available on macOS 12. There is a runtime check for it. On older macOS
versions the flags are stored as a pair of 32bit values. It is slower,
but there are unlikely to be many Cycles users on macOS 11.

Ref #159470

Pull Request: https://projects.blender.org/blender/blender/pulls/162737

+++ intern/cycles/device/cpu/device_impl.cpp
@@ -69,7 +69,7 @@ CPUDevice::~CPUDevice()
   image_info->free();
 }
 
-BVHLayoutMask CPUDevice::get_bvh_layout_mask(uint /*kernel_features*/) const
+BVHLayoutMask CPUDevice::get_bvh_layout_mask(uint64_t /*kernel_features*/) const
 {
   BVHLayoutMask bvh_layout_mask = BVH_LAYOUT_BVH2;
 #ifdef WITH_EMBREE
@@ -361,7 +361,7 @@ void CPUDevice::set_image_cache_func(KernelImageLoadRequestedCPU image_load_requ
   kernel_globals.image_load_requested_cpu = image_load_requested_cpu;
 }
 
-bool CPUDevice::load_kernels(const uint /*kernel_features*/)
+bool CPUDevice::load_kernels(const uint64_t /*kernel_features*/)
 {
   return true;
 }
