+++ intern/cycles/device/oneapi/device_impl.cpp
@@ -173,7 +173,8 @@ bool OneapiDevice::check_peer_access(Device * /*peer_device*/)
   return false;
 }
 
-bool OneapiDevice::can_use_hardware_raytracing_for_features(const uint requested_features) const
+bool OneapiDevice::can_use_hardware_raytracing_for_features(
+    const uint64_t requested_features) const
 {
   /* MNEE and Ray-trace kernels work correctly with Hardware Ray-tracing starting with Embree 4.1.
    */
@@ -185,7 +186,7 @@ bool OneapiDevice::can_use_hardware_raytracing_for_features(const uint requested
 #  endif
 }
 
-BVHLayoutMask OneapiDevice::get_bvh_layout_mask(const uint requested_features) const
+BVHLayoutMask OneapiDevice::get_bvh_layout_mask(const uint64_t requested_features) const
 {
   return (use_hardware_raytracing &&
           can_use_hardware_raytracing_for_features(requested_features)) ?
@@ -251,7 +252,7 @@ size_t OneapiDevice::get_free_mem() const
   return 0;
 }
 
-bool OneapiDevice::load_kernels(const uint requested_features)
+bool OneapiDevice::load_kernels(const uint64_t requested_features)
 {
   assert(device_queue_);
 
@@ -295,7 +296,7 @@ bool OneapiDevice::load_kernels(const uint requested_features)
   return is_finished_ok;
 }
 
-void OneapiDevice::reserve_private_memory(const uint kernel_features)
+void OneapiDevice::reserve_private_memory(const uint64_t kernel_features)
 {
   size_t free_before = get_free_mem();
