+++ intern/cycles/kernel/svm/bevel.h
@@ -290,7 +290,7 @@ ccl_device float3 svm_bevel(
   return is_zero(N) ? sd->N : (sd->runtime_flag & SR_BACKFACING) ? -N : N;
 }
 
-template<uint node_feature_mask, typename ConstIntegratorGenericState>
+template<uint64_t node_feature_mask, typename ConstIntegratorGenericState>
 #  if defined(__KERNEL_OPTIX__)
 ccl_device_inline
 #  else
